% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq03buildW2Mat.R
\name{eq03}
\alias{eq03}
\title{Equation 3: Make northward migration survival matrix (S.m)}
\usage{
eq03(S.m.mg = 0.75, S.m.mp = 0.75, S.m.fg = 0.75, S.m.fp = 0.75)
}
\arguments{
\item{S.m.mg}{Survival (S.) during migration (.m.) of males (.m) originating from good (g) habitat}

\item{S.m.mp}{Survival on migration of males originating from poor (p) habitat.}

\item{S.m.fg}{Migration survival of female birds (f) from good habiatat}

\item{S.m.fp}{Migration survival of females from poor habitat}
}
\description{
This function produces a diagonal matrix of northward (spring)
migration (.m) survival rates
}
\details{
Previous alias Fx.make.spring.mig.surv.eq3
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds.
Johns Hopkins University Press, Baltimore.
}
