% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq04calcBscalar.R
\name{eq04}
\alias{eq04}
\title{Equation 4: Allocate breeding females to Source habitat (B.fc)}
\usage{
eq04(W2, K.bc)
}
\arguments{
\item{W2}{Population vector produced by W0\emph{W1}W2 (?)}

\item{K.bc}{carrying capacity of b.reeding grounds sour.c.e habitat}
}
\value{
xxx xxxx
}
\description{
This function outputs B.fc, the number of females
high quality "source" habitat.
}
\details{
This function compares total female poplatuion size to K.bc
the souce carrying capacity; if total popualtion of females < K.bc, all females
go to the source. If fg + fp > K.bc, then it gets filled up.

The output of this function is used in equation 17.

Previous alias: F.2.source.eq4.c

Part of summer female dynamics
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds.
Johns Hopkins University Press, Baltimore.
}
