% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq04_Bfc.R
\name{eq04_Bfc}
\alias{eq04_Bfc}
\title{Equation 4: Allocate breeding females to Source habitat (B.fc)}
\usage{
eq04_Bfc(W2, K.bc)
}
\arguments{
\item{W2}{Population vector produced after winter and northward migration survival.}

\item{K.bc}{carrying capacity of breeding ground (.k_) source (._c) habitat}
}
\value{
B.fc Abundance ...
}
\description{
This function outputs \code{B.fc}, the number of females that acquire
high quality "source" (._c) habitat during the breeding season
}
\details{
This function compares total female population size after migration (W2.fg + W2.fp)
to K.bc the souce carrying capacity \code{K.bc}; if the total population of
females < K.bc, all females acquire habitat in the source.
If W2.fg + W2.fp > K.bc, then the source habitat gets filled up completely.  Any additinal
females will be allocated to the sink (._k) habitat in a subsequent step.

The output of this function is used in equation 17.

Previous alias: F.2.source.eq4.c

Part of summer female dynamics
}
\examples{
# Set parameters:
## Create named population W2 vector using the eq01buildW0vect()
W2 <- eq01buildW0vect(100,10,100,10)

W2

#Breeding habitat carrying capacity
K.bc.110 <- 110
K.bc.100 <- 100
K.bc.10  <- 10

eq04_Bfc(W2, K.bc.110)
eq04_Bfc(W2, K.bc.100)
eq04_Bfc(W2, K.bc.10)

}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds.
Johns Hopkins University Press, Baltimore.
}
