% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq06calcBscalar.R
\name{eq06}
\alias{eq06}
\title{Equation 6: Allocate breeding males to source habitat.}
\usage{
eq06(W2, K.bc)
}
\arguments{
\item{W2}{Population vector produced by eq02}

\item{K.bc}{carrying capacity...}
}
\value{
xxx xxxx
}
\description{
Equation 6: Allocate breeding males to source habitat.
}
\details{
Summer male (M) dynamces
males acquiring  source habitat
I believe this allows males to be holding territories
w/o pairing w/ a female

Aliased as M.2.source.eq6.c()
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds.
Johns Hopkins University Press, Baltimore.
}
