% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq06_Bmc.R
\name{eq06_Bmc}
\alias{eq06_Bmc}
\title{Equation 6: Allocate breeding males to source habitat (B.mc)}
\usage{
eq06_Bmc(W2, K.bc)
}
\arguments{
\item{W2}{Population vector produced by eq02buildW1Mat()}

\item{K.bc}{Carrying capacity (K) during breeding season (b) of source (c) habitat}
}
\value{
B.mc Total number of males allocated to source habitat.
}
\description{
I believe this allows males to be holding territories without pairing with a female.
}
\details{
TODO(): Check this
}
\examples{
# Standard example
## Set population vector and carrying capacities
W2. <- c(10, 10,10, 10)
W2. <- setNames(W2., c("mg","fg","mp","fp"))
K.bc. <- 10
eq06_Bmc(W2 = W2., K.bc = K.bc.)

}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds.  Birds of two worlds. Johns Hopkins University Press, Baltimore.
}
