% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq08calcBscalar.R
\name{eq08calcScalar}
\alias{eq08calcScalar}
\title{Equation 8: Assigning breeding males to "drain" status}
\usage{
eq08calcScalar(W2, K.bc, B.fk)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{source carrying capacity}

\item{B.fk}{females available in sink}
}
\value{
xxx xxxx
}
\description{
Equation 8: Assigning breeding males to "drain" status
}
\details{
SUMMER MALE (M) DYNAMICS - drain males
Aliased as M.2.drain.eq8()
Drain males:: Males that don't find a territory
AND don't find a mate (for those in sink)
become "drain" males (floaters)
W2\link{"mg"}+W2\link{"mp"} = total male population
minus those that ended up in source habitat (K.bc)
minus those that paired w/female in sink habitat (B.fk)

Aliased previously as M.2.drain.eq8()
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds o
}
