% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq09calcPscalar.R
\name{eq09calcScalar}
\alias{eq09calcScalar}
\title{Equation 9: Pair formation in source habitat between birds both from
good habitat}
\usage{
eq09calcScalar(W2, K.bc, B.mc, B.fc)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{source carrying capacity (?)  original file said "sink popualtion size" in comments}

\item{B.mc}{females available in sink}

\item{B.fc}{...}
}
\value{
xxx, the xxx
}
\description{
Equation 9: Pair formation in source habitat between birds both from
good habitat
}
\details{
In the SOURCE habitat, how many PAIRS are good-good
(both male and female originate from good wintering habitat)
"of all PAIRS in source habitat, the proportion composed of both
a male and female FROM good habitat is
This function determines how many females from good winter habitat
pair w/ males from good habitat
1)If W2\link{"mg"} > K.bc  (number of males from good habitat > source K)
AND W2\link{"fg"} > K.bc (number of females from good > source K)
Then all of the pairs on the source habitat will be good-good pairs
That is, if there are enough males and femaels to fill up the source
then the pairing rate will be 100%
This would only occur when there are many birds
2)Otherwise, the ratio of good-good pairs will be determined
by the ratio of
2a)the number of males or females from good habitat, whichever is less
2b)and the number of males or females on source habitat, whichever is less
(?)

B.fc should usually be lower than B.mc b/c females have higher mort and/or are forced
by competition into poor habitat in the winter, resulting in carry over effects

Possible ratios for the 2nd condition
W2.mg/B.mc = males from good winter / males settled in source
W2.fg/B.mc = females from good winter / males settled in source
W2.mg/B.fc = males from good winter / females settled in source
W2.fg/B.fc = females from good winter / females settled in source
The ratio r is tricky.  It is determined by which sex is limiting in terms of
input from good winter habitat (numerator)
and which sex is limiting in terms of total abundance in the breeding habitat

Aliased as pairing.eq9.P.c.gg() in previous version of coad
}
