% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq10calcPscalar.R
\name{eq10}
\alias{eq10}
\title{Equation 10:}
\usage{
eq10(W2, K.bc, B.fc, B.mc)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{females on sour.c.habitat}

\item{B.fc}{females available in sink}

\item{B.mc}{males on sour.c. habitat}
}
\value{
xxx xxxx
}
\description{
Equation 10:
}
\details{
SUMMER MALE (M) DYNAMICS - pairing dynamics
proportion of pairs on source habitat
composed of MALES   from GOOD and
females from poor
1)If the number of males from good winter habitat is greater than the
the number of females from good habitat
AND the number of females from good is less than the total K for sour.c.e habitat
Then the number of good-poor pairings is determined by...
2)other, the number of good-poor pairings is zero (all pairings are good-good)

Aliased prevously as pairing.eq10.P.c.gp
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds o
}
