% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq10_Pcgp.R
\name{eq10_Pcgp}
\alias{eq10_Pcgp}
\title{Equation 10: proportion males from good winter habitat paired with females from poor winter habitat (P.cgp)}
\usage{
eq10_Pcgp(W2, K.bc, B.mc, B.fc, ...)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{Carrying capacity (K) during breeding season (b) of source (c) habitat}

\item{B.mc}{females available in sink (typo; males?, source?)}

\item{B.fc}{males on sour.c. habitat (typo, females?)}

\item{...}{Additional parameters}
}
\value{
P.cgp proportion males from good winter habitat paired with females from poor winter habitat
}
\description{
The logic of this function is:
\enumerate{
\item IF (the number of males from good winter habitat) > (the number of females from good habitat)
AND (the number of females from good) <  (the total K for source habitat)
THEN the number of good-poor pairings is determined by TODO()...
\item ELSE, the number of good-poor pairings is 0 (all pairings are good-good)
}
}
\details{
Note that in the original paper, both equations 10 and 11 have the same subscripts of P.cgp, while in equation 12 it is implied they have different subscripts.  It appears that the 2nd subscript should indicate the winter habitat for males and the 3rd should indicate the winter habitat for females.  Therefore equation 10 in the original paper remains as P.cgp (P.source.male-good.female-poor) and equation 11 should be changed to P.c.pg (P.source.male-poor.female-good)
}
\examples{
# Trivial example
W2. <- c(10,10,10,10)
names(W2.) <- c("mg","fg","mp","fp")
eq10_Pcgp(W2 = W2., K.bc = 0, K.bk = 10, B.mc = 5, B.fc =5)



}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds.  Birds of two worlds. Johns Hopkins University Press, Baltimore.
}
