% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq11calcPscalar.R
\name{eq11}
\alias{eq11}
\title{Equation 11:}
\usage{
eq11(W2, B.mc, B.fc, K.bc)
}
\arguments{
\item{W2}{population vector; indicats where birds are coming from}

\item{B.mc}{males already in source}

\item{B.fc}{females alreadiy in source}

\item{K.bc}{females  sour.c.habitat carrying capacity}
}
\value{
xxx xxxx
}
\description{
Equation 11:
}
\details{
SUMMER MALE (M) DYNAMICS
Aliased as pairing.eq11.P.cpg()
proportion of poor males mated w/ "good" female
1)If males that wintered in good habitat > females that wintered in good habitat
BUT males that wintered in good less than source K
Then the proportion of poor-good pairings is
a function of
Otherwise, there are no poor-good pairings

Previously aliased as pairing.eq11.P.cpg
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds o
}
