% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq12calcPscalar.R
\name{eq12}
\alias{eq12}
\title{Equation 12}
\usage{
eq12(P.cgg, P.cgp, P.cpg)
}
\arguments{
\item{P.cgg}{proportion...}

\item{P.cgp}{proportion...}

\item{P.cpg}{proportion...}
}
\value{
P.cpp
}
\description{
Equation 12
}
\details{
SUMMER MALE (M) DYNAMICS - pairing
proportion composed of a male and female both from poor habitat
This is calcualted by subtraction

Previously aliased as pairing.eq12.P.cpp()
}
