% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq14calcPscalar.R
\name{eq14}
\alias{eq14}
\title{Equation 14: proportion in sink habitat, good-poor pairs}
\usage{
eq14(W2, K.bc, K.bk, B.mk, B.fk)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{carrying capacity ...}

\item{K.bk}{carrying capacity ...}

\item{B.mk}{...}

\item{B.fk}{...}
}
\description{
Equation 14: proportion in sink habitat, good-poor pairs
}
\details{
Previously aliased as pairing.eq14.P.kgp()
}
