% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq15calcPscalar.R
\name{eq15}
\alias{eq15}
\title{Equation 15: Proportion in sink poor-poor}
\usage{
eq15(W2, K.bc, K.bk, B.mk, B.fk)
}
\arguments{
\item{W2}{Popualtion vector}

\item{K.bc}{carrying capacity}

\item{K.bk}{carrying capacity}

\item{B.mk}{...}

\item{B.fk}{...}
}
\description{
Equation 15: Proportion in sink poor-poor
}
\details{
Aliased as pairing.eq15.P.kpg()
}
