% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq16calcPscalar.R
\name{eq16}
\alias{eq16}
\title{Equation 16:}
\usage{
eq16(P.kgg, P.kgp, P.kpg, B.mk, B.fk)
}
\arguments{
\item{P.kgg}{...}

\item{P.kgp}{...}

\item{P.kpg}{...}

\item{B.mk}{...}

\item{B.fk}{...}
}
\description{
Equation 16:
}
\details{
Proportion in sink composed of poor-poor

Previously aliased as pairing.eq16.P.kpp
}
