% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq18BbuildPmat.R
\name{eq18BbuildPmat}
\alias{eq18BbuildPmat}
\title{Equation 18: P matrix - Build 2 x 8 matrix of the proporations in each habitat combo (P.all)}
\usage{
eq18BbuildPmat(
  P.cgg = 1,
  P.cgp = 0,
  P.cpg = 0,
  P.cpp = 0,
  P.kgg = 1,
  P.kgp = 0,
  P.kpg = 0,
  P.kpp = 0
)
}
\arguments{
\item{P.cgg}{From equation 9: the proportion of pairings in the source breeding habitat between male and females which both winter in the good habitat}

\item{P.cgp}{From equation 10: proportion males from good winter habitat paired with females from poor winter habitat}

\item{P.cpg}{From equation 11: the proportion of pairings in the source habitat (c) made up of males from poor winter habitat (p) and females from good winter habitat (g).}

\item{P.cpp}{From equation 12: Proportion of prairs between a male and female both from poor habitat}

\item{P.kgg}{From equation 13:Proportion of pairs in sink (k) habitat where both male and female originated from good winter habitat (gg)}

\item{P.kgp}{From equation 14: Proportion of pairs in sink (k) which are good-poor pairings4}

\item{P.kpg}{From Equation 15: Proportion of pairs in sink (k) habitat where both male originates from poor (p) and female originated from good (g) winter habitat}

\item{P.kpp}{From: Equation 16: Proportion of pairs in sink (k) which are poor-poor pairings}
}
\value{
P.all 2 x 8 matrix of the proporiton in each habitat
}
\description{
"Average fecunidty for pairs in source and sink habitat" is a function of how the proportion of pairs in that habitat that are good-good, good-poor etc (Runge and Marra 2004, pg TOD())
}
\examples{
# Shows contents of the vector
eq18BbuildPmat()

}
