% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq18buildPmat.R
\name{eq18buildPmat}
\alias{eq18buildPmat}
\title{Equation 18: Build 2 x 8 matrix of the proporations in each habitat combo}
\usage{
eq18buildPmat(P.cgg, P.cgp, P.cpg, P.cpp, P.kgg, P.kgp, P.kpg, P.kpp)
}
\arguments{
\item{P.cgg}{proportion ... Calculated by equation 9}

\item{P.cgp}{proportion ... Calcualted by equation 10}

\item{P.cpg}{proportion ... Calcualted by equation 11 (?)}

\item{P.cpp}{proportion ... Calcluated by subtraction using equation 12}

\item{P.kgg}{proportion ... Calcualted by eqation 13}

\item{P.kgp}{proportion ... Calcualted by eqation 14}

\item{P.kpg}{proportion ... Calcualted by eqation 15}

\item{P.kpp}{proportion ... Calcualted by eqation 16}
}
\value{
P.all 2 x 8 matrix of the proporiton in each habitat
}
\description{
Average fecunidty for pairs in source and sink habitat"
is a function of how the proportion of pairs in that
habitat that are good-good, good-poor etc
}
\details{
Previously aliased as Fx.make.P.matrix.eq18
}
