% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq18buildRvect.R
\name{eq18buildRmat}
\alias{eq18buildRmat}
\title{Equation 18: R vector - build vector of reproduction/fecunity values}
\usage{
eq18buildRmat(R.base.rate, R.hab.effect, co)
}
\arguments{
\item{R.base.rate}{Base rate of...}

\item{R.hab.effect}{Effect of breeding in poor habitat; in table 28.3 base fec is 1.8, while in poor habitat it is 0.9}

\item{co}{strength of carry over effect}
}
\value{
R.all All reproductive values (?)
}
\description{
fx.make.R.all.eq18()
Fx.make.P.matrix.eq18()
LOAD EQUATION 18b
}
\details{
REPRODUCTION
"average fecunidty for pairs in source and sink habitat"
is a function of how the proportion of pairs in that
habitat that are good-good, good-poor etc

These are fixed for the entire duratin of the study

Calculations are found in Table 28.3.
}
