% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq19AbuildFmat.R
\name{eq19AbuildFmat}
\alias{eq19AbuildFmat}
\title{Equation 19: Matrix of ... (f.mat)}
\usage{
eq19AbuildFmat(f = 0.5)
}
\arguments{
\item{f}{Fecundity?}
}
\value{
f.mat matrix of xxx
}
\description{
sex ratio matrix (equation 19)
}
\examples{
# Sex ratio matrix
eq19AbuildFmat()
}
