% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq19BbuildMinMat.R
\name{eq19BbuildMinMat}
\alias{eq19BbuildMinMat}
\title{Equation 19: Build 2 x 2 Diagonal matrix based in minimumbs of B.mc vs B.fc and B.mk vs. B.fk}
\usage{
eq19BbuildMinMat(B.mc, B.fc, B.mk, B.fk)
}
\arguments{
\item{B.mc}{xxx}

\item{B.fc}{xxx}

\item{B.mk}{xxx}

\item{B.fk}{xxx}
}
\value{
min.mat xxx
}
\description{
Fx.make.min.mat.eq19()
}
\details{
Previously aliased as Fx.make.min.mat.eq19
}
