% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq19buildFmat.R
\name{eq19buildFmat}
\alias{eq19buildFmat}
\title{Equation xx Fx.make.F.matrix()}
\usage{
eq19buildFmat(f = 0.5)
}
\arguments{
\item{f}{Fecundity?}
}
\description{
Equation xx Fx.make.F.matrix()
}
\details{
sex ratio matrix (equation 19)

Formerly aliased as Fx.make.F.matrix, make.F.matrix
}
\examples{
# Sex ratio matrix
eq19buildFmat()
}
