% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq19buildMinMat.R
\name{eq19buildMinMat}
\alias{eq19buildMinMat}
\title{Equation 19: Build 2 x 2 Diagonal matrix based in minimumbs of B.mc vs B.fc and B.mk vs. B.fk}
\usage{
eq19buildMinMat(B.mc, B.fc, B.mk, B.fk)
}
\arguments{
\item{B.mc}{...}

\item{B.fc}{...}

\item{B.mk}{...}

\item{B.fk}{...}
}
\value{
min.mat ...
}
\description{
Fx.make.min.mat.eq19()
}
\details{
Previously aliased as Fx.make.min.mat.eq19
}
