% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq20_build_Sb_mat.R
\name{eq20_build_Sb_mat}
\alias{eq20_build_Sb_mat}
\title{Equation 20: BREEDING season mortality}
\usage{
eq20_build_Sb_mat(S.b.mc, S.b.mk, S.b.md, S.b.fc, S.b.fk)
}
\arguments{
\item{S.b.mc}{survival...}

\item{S.b.mk}{survival...}

\item{S.b.md}{survival...}

\item{S.b.fc}{survival...}

\item{S.b.fk}{survival...}
}
\description{
Equation 20: BREEDING season mortality
}
\details{
"adult birds experience both sex- and habitat-specific mortality
over the breeding season."

Previously aliased as Fx.make.breeding.s.matrix.eq20
}
