% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq20buildSbmat.R
\name{eq20buildSbmat}
\alias{eq20buildSbmat}
\title{Equation 20: Breeding-season mortality matrix (S.b)}
\usage{
eq20buildSbmat(
  S.b.mc = 0.95,
  S.b.mk = 0.85,
  S.b.md = 0.8,
  S.b.fc = 0.95,
  S.b.fk = 0.85
)
}
\arguments{
\item{S.b.mc}{Survival (S) during summer breeding season (b) of adult males (m) who hold territories on source (c) habitat.}

\item{S.b.mk}{Survival (S) during summer breeding season (b) of adult males (m) who are nesting in sink (k) habitat.}

\item{S.b.md}{Survival (S) during summer breeding season (b) of adult males (m) who have "drain" (d) status and do not hold a territory or have a make (aka unpaired males, floaters)}

\item{S.b.fc}{Survival (S) during summer breeding season (b) of adult females (f) who are nesting source (c) habitat.}

\item{S.b.fk}{Survival (S) during summer breeding season (b) of adult females (f) who are nesting in on sink (k) habitat.}
}
\value{
S.b Diagonal matrix of breeding-season survival probabilities
}
\description{
"Adult birds experience both sex- and habitat-specific mortality over the breeding season" (Runge and Marra 2004, page TODO())
}
\examples{
eq20buildSbmat()

}
