% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq21buildMat.R
\name{eq21}
\alias{eq21}
\title{Equation 21 - MIGRATION Mortality of ADULTS
"Mort during mig depends upon the sex of the bird and breeding habitat it used}
\usage{
eq21(S.f.mc, S.f.mk, S.f.md, S.f.fc, S.f.fk)
}
\arguments{
\item{S.f.mc}{survival...}

\item{S.f.mk}{survival...}

\item{S.f.md}{survival...}

\item{S.f.fc}{survival...}

\item{S.f.fk}{survival...}
}
\description{
Aliased previously Fx.make.fall.adult.s.matrix.eq21
}
