% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq21buildSfmat.R
\name{eq21buildSfmat}
\alias{eq21buildSfmat}
\title{Equation 21: Migration mortality matrix for adults (S.f)}
\usage{
eq21buildSfmat(
  S.f.mc = 0.8,
  S.f.mk = 0.75,
  S.f.md = 0.8,
  S.f.fc = 0.8,
  S.f.fk = 0.75
)
}
\arguments{
\item{S.f.mc}{Survival (S) during fall migration (f) of males (m) originating from source (c) habitat.}

\item{S.f.mk}{Survival (S) during fall migration (f) of males (m) originating from sink (k) habitat.}

\item{S.f.md}{Survival (S) during vall migration (f) of males (m) who had "drain" (d) status, aka floaters or unpaired males.}

\item{S.f.fc}{Survival (S) during fall migration (f) of females (f) originating from source (c) habitat.}

\item{S.f.fk}{Survival (S) during fall migration (f) of females (f) originating from sink (k) habitat.}

\item{S.f}{Diagonal matrix of fall survival probabilities for adults}
}
\description{
"Mortality during migration depends upon the sex of the bird and breeding habitat it used" (Runge and Marra 2004, page TOD())
}
\examples{
eq21buildSfmat()

}
