% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq22_build_Sy_mat.R
\name{eq22_build_Sy_mat}
\alias{eq22_build_Sy_mat}
\title{Equation 22: Mortality during migration of OFFSPRING}
\usage{
eq22_build_Sy_mat(S.y.mc, S.y.mk, S.y.fc, S.y.fk)
}
\arguments{
\item{S.y.mc}{survival...}

\item{S.y.mk}{survival...}

\item{S.y.fc}{survival...}

\item{S.y.fk}{survival...}
}
\description{
Equation 22: Mortality during migration of OFFSPRING
}
\details{
Previously aliased as Fx.make.fall.adult.s.matrix.eq22
}
