% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq22buildSymat.R
\name{eq22buildSymat}
\alias{eq22buildSymat}
\title{Equation 22: Mortality during migration of young/offspring (S.y)}
\usage{
eq22buildSymat(S.y.mc = 0.8, S.y.mk = 0.75, S.y.fc = 0.8, S.y.fk = 0.75)
}
\arguments{
\item{S.y.mc}{Survival of young (y; aka juveniles) males (m) originating from source (c) habitat}

\item{S.y.mk}{Survival of young (y) males (m) originating from sink (k) habitat}

\item{S.y.fc}{Survival of young (y) females (f) originating from source (c) habitat}

\item{S.y.fk}{Survival of young (y) females (f) originating from sink (k) habitat}
}
\value{
S.y Matrix of survival probabilities for young.
}
\description{
Equation 22: Mortality during migration of young/offspring (S.y)
}
\examples{

# Return default values
eq22buildSymat()

}
