% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq23_stack_Ai.R
\name{eq23_stack_Ai}
\alias{eq23_stack_Ai}
\title{Equation 23: Build the A matrix (not implemented)}
\usage{
eq23_stack_Ai(B2, Y2)
}
\arguments{
\item{B2}{Adults that survived breeding and fall migration}

\item{Y2}{Young (offspring) that survived breeding and fall migration}
}
\description{
Equation 23: Build the A matrix (not implemented)
}
\details{
...
}
