% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq23buildAvect.R
\name{eq23buildAvect}
\alias{eq23buildAvect}
\title{Equation 23: Build A matrix (not implemented)}
\usage{
eq23buildAvect(B2, Y2)
}
\arguments{
\item{B2}{Adults that survived breeding and fall migration}

\item{Y2}{Young (offspring) that survived breeding and fall migration}
}
\description{
Equation 23: Build A matrix (not implemented)
}
