% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq24_competition.R
\name{eq24_competition}
\alias{eq24_competition}
\title{Equation 24: competition for winter habitats}
\usage{
eq24_competition(A.i.j, y.i, K.wg.open.j, ...)
}
\arguments{
\item{A.i.j}{xxx}

\item{y.i}{xxx}

\item{K.wg.open.j}{xxx}

\item{...}{xxx}
}
\value{
A.i.settled.raw.j
}
\description{
Equation 24: competition for winter habitats
}
