% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq24_make_gamma_vec.R
\name{eq24_make_gamma_vec}
\alias{eq24_make_gamma_vec}
\title{Equation 24: Make relative competitivness parameters for winter habitat acquition}
\usage{
eq24_make_gamma_vec(gamma.base = 2)
}
\arguments{
\item{gamma.base}{Base competition value of females relative to males.   Higher values make females LESS competitive.}
}
\value{
gamma.i Vector of competiveness values for each class of birds
}
\description{
Competitive ability depends on "an intrsince age-,sex- and condition(habitat)-specific competitive factor gamma and the number of birds in each class" (Runge and Mara, 2004; page 380, column 2).  THis function calculates gammas for each class of birds based on gamma.base provided in function call.  Currently competitiveness for winter habitat is primarily determined by age and the habitat a bird originated (source vs. sink).  Changing gamma.base decreases the competitiveness of females relative to males.
}
\details{
Note: breeding K is based on pairs,winter K is based on individuals.
}
\examples{

# Default relative competitiveness is that males = females
gamma.1 <- eq24_make_gamma_vec()

# Make females more competitive than males
gamma.1.2 <- eq24_make_gamma_vec(gamma.base = 1.2)

data.frame(gamma.1, gamma.1.2)

}
