% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq24buildGammaVect.R
\name{eq24buildGammaVect}
\alias{eq24buildGammaVect}
\title{Equation 24 - Competition parameters (gamma)}
\usage{
eq24buildGammaVect(gamma.base)
}
\arguments{
\item{gamma.base}{Competition}
}
\description{
Calculates gammas based on gamma.base provided in function call
}
\details{
NB: breeding K is based on pairs,
winter K is based on individuals!
competition ability depends on "an intrsince age-,sex-
and condition(habitat)-specific competitive factor
gamma and the number of birds in each class"

gamma represents "an instrinsic age-,sex,- and condition (habitat)-specific...factor" pg380 col 2

previously aliased as fx.make.gamma.i.eq24
}
