% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq24compLoop.R
\name{eq24compLoop}
\alias{eq24compLoop}
\title{Equation 24: competition - Allocation bird to high quality winter habitiat}
\usage{
eq24compLoop(K.wg, A.i, gamma.i)
}
\arguments{
\item{K.wg}{Carrying capacity xxx}

\item{A.i}{xxx}

\item{gamma.i}{competition xxx}
}
\description{
Equation 24: competition - Allocation bird to high quality winter habitiat
}
\details{
A.i = initial population of vector when competition process starts
A.i.i = vector of individuals that haven't been allocated yet
A.G.i = vector of weights of individuals in A.i.i; this is constrainted so that A.G.i cannot be greater thatn A.i.i
after constrain implemented, this is the vector of indiviuals alloacted
to different habitat qualities
A.storage.i = vector to hold output of each loop
A.storage = object to hold entire output of competition process?
}
