% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq27_post_comp_pooling.R
\name{eq27_post_comp_pooling}
\alias{eq27_post_comp_pooling}
\title{Equation 27:}
\usage{
eq27_post_comp_pooling(A.i.G, A.i.P)
}
\arguments{
\item{A.i.G}{xxx}

\item{A.i.P}{xxx}
}
\value{
W.list xxx
}
\description{
Equation 27:
}
