% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_B0.R
\name{error_check_B0}
\alias{error_check_B0}
\title{QA/QC: Check for errors in B0 (vector of abundances after breeding habiatat acquisition)}
\usage{
error_check_B0(B0, W2, check.errors.in, i, ...)
}
\arguments{
\item{B0}{5 x 1 vector representing population state after breeding habitat acqusitions}

\item{W2}{4 x 1 vectir representing populati state after northward spring migration}

\item{check.errors.in}{vector of model components to check for errors during a given run}

\item{i}{Current iteration of runFAC()}

\item{...}{additional parameters}
}
\description{
QA/QC: Check for errors in B0 (vector of abundances after breeding habiatat acquisition)
}
\examples{
B0 <- c(10,20,30,40,50)
W2 <- c(10,20,30,40)
error_check_B0(B0 = B0,
W2 = W2,
check.errors.in = "B0",
i = 1)

}
