% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_param_set.R
\name{error_check_param_set}
\alias{error_check_param_set}
\title{Quality control on initial parameters entering runFAC()}
\usage{
error_check_param_set(param.set)
}
\description{
Quality control on initial parameters entering runFAC()
}
