% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_Pxxx.R
\name{error_check_Pxxx}
\alias{error_check_Pxxx}
\title{QA/QC: Check for errors in pairing equations}
\usage{
error_check_Pxxx(P.xxx.name, P.xxx.value, P.xxx.eq.name = NA, i = NA,
  ...)
}
\arguments{
\item{P.xxx.name}{name of scalar to check}

\item{P.xxx.value}{value for a given P.xxx}

\item{P.xxx.eq.name}{full name of equation}

\item{i}{Current iteration of runFAC()}

\item{...}{additional paramters}
}
\description{
QA/QC: Check for errors in pairing equations
}
\examples{
P.xxx.names <- c("P.cgg", "P.cgp","P.cpg", "P.cpp","P.kgg", "P.kgp","P.kpg", "P.kpp")
P.xxx.values <- runif(min = -0.8,max = 1.2,n= length(P.xxx.names))
P.xxx.values[3] <- NaN

error_check_Pxxx(P.xxx.name = P.xxx.names[3],
                 P.xxx.value = P.xxx.values[3])

for(i in 1:length(P.xxx.names)){
   error_check_Pxxx(P.xxx.name = P.xxx.names[i],
                 P.xxx.value = P.xxx.values[i])}
}
