% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_summer.R
\name{error_check_summer}
\alias{error_check_summer}
\title{QA/QC: Check output of breeding season processes for errors}
\usage{
error_check_summer(B0, W2, P.breeding.pair.results, check.errors.in, i)
}
\arguments{
\item{B0}{xxx}

\item{W2}{xxx}

\item{P.breeding.pair.results}{xxxx}

\item{check.errors.in}{what to check errors in}

\item{i}{index}
}
\description{
QA/QC: Check output of breeding season processes for errors
}
