% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_Y1_sex_ratio.R
\name{error_check_Y1_sex_ratio}
\alias{error_check_Y1_sex_ratio}
\title{Check offspring sex ratio}
\usage{
error_check_Y1_sex_ratio(Y1, i, ...)
}
\arguments{
\item{Y1}{Vector of offspring abundance}

\item{i}{current iteration}

\item{...}{...}
}
\description{
Check offspring sex ratio
}
