% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redstart-package.R
\docType{package}
\name{redstart-package}
\alias{redstart}
\alias{redstart-package}
\title{redstart: An R implementation and extention of full-annual cycle (FAC) model of Runge & Marra 2004}
\description{
redstart is an R implementation and extension of the full annual cycle (FAC) model for migratory birds developed by Michael Runge and Peter Marra and published in Birds of Two Worlds.  The original model is a periodic, 2-sex, density-dependent matrix model that incorporates seasonal dynamics, carryover effects, and two classes of habitat on both the summer breeding grounds.  redstart is an R implementation of the model that incorporates Monte Carlo and sensitivity analyses.  Thanks goes to M. Runge for graciously answering questions and providing Matlab code snippets.
}
\author{
\strong{Maintainer}: Nathan Brouwer \email{brouwern@gmail.com}

}
\keyword{internal}
