% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FAC_df.R
\name{make_FAC_df}
\alias{make_FAC_df}
\title{Create FAC output dataframe}
\usage{
make_FAC_df(iterations = 300)
}
\arguments{
\item{iterations}{Number of iterations to run basic FAC model. Defaults to 300.}
}
\value{
out.df A dataframe to hold output of each iteration from a single run of the model.
}
\description{
Creates a dataframe to hold output from a single run
of a Runge & Marra (2004) FAC model Model is run over a certain number of
iterations (reps) inorder to reach equilibrium. The main output of a full run
of model is the equilibrium population size.
}
