% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_multiFAC_df.R
\name{make_multiFAC_df}
\alias{make_multiFAC_df}
\title{Make empty dataframe for holding output of runMultiFAC(). which runs the model accross a range of paramers}
\usage{
make_multiFAC_df(params.use = param_grid())
}
\arguments{
\item{params.use}{Parameter values to use}
}
\description{
Make empty dataframe for holding output of runMultiFAC(). which runs the model accross a range of paramers
}
