% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParSeq.R
\name{makeParamSeq}
\alias{makeParamSeq}
\title{Make parameter range for looping over multiple values of a parameter}
\usage{
makeParamSeq(rangeParam, var.length)
}
\arguments{
\item{rangeParam}{...}

\item{var.length}{...}
}
\value{
param.seq Dataframe with 2 columns, minimum and maxium values for each parameter
}
\description{
Make parameter range for looping over multiple values of a parameter
}
