% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParCombo.R
\name{makeParCombos}
\alias{makeParCombos}
\title{Make combinations of parameters for parameter matrix}
\usage{
makeParCombos(paramRanges = makeParamRanges.df(), var.length = 10)
}
\arguments{
\item{paramRanges}{Input range (min, max) for each paramters}

\item{var.length}{Step size (?)}
}
\value{
ParamCombos.df Dataframe of parameter combinations
}
\description{
This takes a df of parameter ranges generated by paramRanges.df()
and applied makeParamSeq() to turn any that vary into a sequence spanning
the range.
These sequences (if any) are then turned into a matrix of all possible values
if none vary, then a single row of data is returned
}
\details{
Formerly aliased as makeParamCombos.df
}
