% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOut_df.R
\name{makeSingleRunOutDf}
\alias{makeSingleRunOutDf}
\title{Create FAC output dataframe function}
\usage{
makeSingleRunOutDf(reps = 300)
}
\arguments{
\item{reps}{Number of reps to run basic FAC model. Defaults to 300.}
}
\value{
out.df A dataframe to hold output of each iteration from a single run of the model.
}
\description{
Creates a dataframe to hold output from a single run
of a Runge & Marra (2004) FAC model Model is run over a certain number of
iterations (reps) inorder to reach equilibrium. The main output of a full run
of model is the equilibrium population size.
}
\details{
Formerly aliased as fx.create.out.df
}
