% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_ranges.R
\name{param_ranges}
\alias{param_ranges}
\title{Set minimum and maximum values for all parameters.}
\usage{
param_ranges(scenario = NA, figure = NA, gamma. = c(5, 5),
  c. = c(1, 1), K.bc. = c(1000, 1000), K.bk. = c(10000, 10000),
  K.wg. = c(900, 900), S.w.mg. = c(0.8, 0.8), S.w.mp. = c(0.8, 0.8),
  S.w.fg. = c(0.8, 0.8), S.w.fp. = c(0.8, 0.8), S.m.mg. = c(0.9,
  0.9), S.m.mp. = c(0.8, 0.8), S.m.fg. = c(0.9, 0.9),
  S.m.fp. = c(0.8, 0.8), R.base.rate. = c(1.8, 1.8),
  R.hab.effect. = c(0.5, 0.5), co. = c(1, 1), f. = c(0.5, 0.5),
  S.b.mc. = c(0.95, 0.95), S.b.mk. = c(0.85, 0.85), S.b.md. = c(0.8,
  0.8), S.b.fc. = c(0.95, 0.95), S.b.fk. = c(0.85, 0.85),
  S.f.mc. = c(0.8, 0.8), S.f.mk. = c(0.75, 0.75), S.f.md. = c(0.8,
  0.8), S.f.fc. = c(0.8, 0.8), S.f.fk. = c(0.75, 0.75),
  S.y.mc. = c(0.8, 0.8), S.y.mk. = c(0.75, 0.75), S.y.fc. = c(0.8,
  0.8), S.y.fk. = c(0.75, 0.75), verbose = TRUE)
}
\arguments{
\item{scenario}{Different habitat scenarios defined by Runge & Marra 2004.  In particular sets carrying capacity for breeding source habitat (K.bc) and winter good habitat (K.wg)    If set overides other varibles that may have been set using other arguements.}

\item{figure}{Generate figure.  Currently only figures 28.3 and 28.4 are implemented}

\item{gamma.}{Strength of male dominance.}

\item{c.}{Strength of winter to breeding season carry over.}

\item{K.bc.}{Carrying capacity (K) of source (high quality) breeding habitat.}

\item{K.bk.}{Carrying capacity (K) of sink (low quality) breeding habitat.}

\item{K.wg.}{Carrying capacity (K) of winter good (high quality) habitat.}

\item{S.w.mg.}{Survival (S) during winter (.w.) of males (.m) in good habitat (g).  Note that there is no age structure in winter.}

\item{S.w.mp.}{Survival (S) during winter (.w.) of males (.m) in poor habitat (p).}

\item{S.w.fg.}{Survival (S) during winter (.w.) of females (.f) in good habitat (g).}

\item{S.w.fp.}{Survival (S) during winter (.w.) of females (.f) in good habitat (p).}

\item{S.m.mg.}{Survival (S) during spring migration (.m.) of males (.m) originating from good (g) winter habitat.}

\item{S.m.mp.}{Survival (S) during spring migration (.m.) of males (.m) originating from poor (p) winter habitat.}

\item{S.m.fg.}{Survival (S) during spring migration (.m.) of females (.f) originating from good (g) winter habitat.}

\item{S.m.fp.}{Survival (S) during spring migration (.m.) of females (.m) originating from poor (p) winter habitat.}

\item{R.base.rate.}{Fecundity base rate (?)}

\item{R.hab.effect.}{Effect on fecundity of a pair breeding in poor habitat}

\item{co.}{Magnitude of carry over effect; c = 1 is no carry over}

\item{f.}{Sex ratio}

\item{S.b.mc.}{Survival (S) during breeding (.b.) of males (.m) in source (c) habitat (source = good habitat)}

\item{S.b.mk.}{Survival (S) during breeding (.b.) of males (.m) in sink (k) habitat (sink = poor habitat)}

\item{S.b.md.}{Survival (S) during breeding (.b.) of males (.m) with drain (d) status (drain = floater or unpaired)}

\item{S.b.fc.}{Survival (S) during breeding (.b.) of females (.f) in source (c) habitat (source = good habitat)}

\item{S.b.fk.}{Survival (S) during breeding (.b.) of females (.f) in sink (k) habitat (sink = poor habitat)}

\item{S.f.mc.}{Survival (S) during fall (.f.) migration of males (.m) originating from source (c) habitat}

\item{S.f.mk.}{Survival (S) during fall (.f.) migration of males (.m) originating from sink (k) habitat}

\item{S.f.md.}{Survival (S) during fall (.f.) migration of males (.m) with drain (d) status (aka floater, unpaired)}

\item{S.f.fc.}{Survival (S) during fall (.f.) migration of females (.f) originating from source (c) habitat}

\item{S.f.fk.}{Survival (S) during fall (.f.) migration of females (.f) originating from sink (k) habitat}

\item{S.y.mc.}{Survival (S) during fall migration of young (.y.) males (.m) originating from source (c) habitat}

\item{S.y.mk.}{Survival (S) during fall migration of young (.y.) males (.m) originating from sink (k) habitat}

\item{S.y.fc.}{Survival (S) during fall migration of young (.y.) females (.f) originating from source (c) habitat}

\item{S.y.fk.}{Survival (S) during fall migration of young (.y.) females (.f) originating from sink (k) habitat}

\item{verbose}{T/F return messages}
}
\value{
param.ranges 30 x 2 dataframe of the min. and max. parameter values that will be considered.  Often values are fixed and do not vary and so are listed as both min and max.
}
\description{
Generates a dataframe defining the minimum and maximum values of each parameter.
This dataframe of minimums and maximums is then turned into a list if scalers (single numbers) for parameters that do not vary and
and vectors for numbers that do vary by the function param_seqs().  A dataframe of all possible parameter combinatnsion (aka, a parameter grid) is then made with param_grid().
}
\details{
Default values are based on Runage and Marra (2004).  The arguments "scenario" and "figure" can be used to set up the parameters for specific figures from Runge and Marra (2004).
}
\examples{

# Set parameters for the winter limited scenario in original paper (Runge & Marra 2004)
param.ranges <- param_ranges(scenario = "winter.limited")
head(param.ranges)

# Parameters to remake Figure 28.3 in original paper (Runge & Marra 2004)
param.ranges <- param_ranges(figure = 28.3)
param.ranges[c("K.bc","K.wg"), ]

# Setting carrying capacities to 0 causes errors when running model
## param_ranges() therefore automatically changes to prevent this bug
param.ranges <- param_ranges(K.bc. = c(0,100))
param.ranges[c("K.bc","K.wg"), ]

# Set up different scenarios
## this is similar to data for figure 28.5, but without gamma varying
winter.lim <- param_ranges(scenario = "winter")
intermed <- param_ranges(scenario = "intermediate")
summer.lim <- param_ranges(scenario = "summer")

# Show data
## Note only 1st 6 rows shown; delete "[1:6,]" to show all.
data.frame(winter.lim, intermed, summer.lim)[1:6,]

}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds: The ecology & evolution of migration.
Johns Hopkins University Press, Baltimore.
}
