% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_seqs.R
\name{param_seqs}
\alias{param_seqs}
\title{Create sequence of parameter values between minimum and maximum.}
\usage{
param_seqs(param.ranges = param_ranges(), len.out = 10)
}
\arguments{
\item{param.ranges}{minimum and maximum values for all parameters}

\item{len.out}{number of increments to divide range into}
}
\value{
param.seq List with 30 elements, one for each parameter.
}
\description{
Create sequence of parameter values between minimum and maximum.
}
