% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_seqs2.R
\name{param_seqs2}
\alias{param_seqs2}
\title{Create sequence of parameter values between minimum and maximum.}
\usage{
param_seqs2(param.ranges = param_ranges(), len.out = 10)
}
\arguments{
\item{param.ranges}{minimum and maximum values for all parameters, generted by \link{param_ranges}}

\item{len.out}{number of increments to divide range into}
}
\value{
param.seq List with 30 elements, one for each parameter.
}
\description{
Create sequence of parameter values between minimum and maximum.
}
\examples{
#generate min and max values for each parameter used for figure 28.3 in original paper
params <- param_ranges(figure = 28.3)

# Generate parameter sequence
params.seqs <- param_seqs(param.ranges = params)


## Most are constants; 3rd and 5th element vary accross a range of values
params.seqs$K.wg



}
