% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{param_set}
\alias{param_set}
\title{Make paramater ranges}
\usage{
param_set(scenario = NA, figure = NA, gamma. = 5, c. = 1,
  K.bc. = 1000, K.bk. = 10000, K.wg. = 900, S.w.mg. = 0.8,
  S.w.mp. = 0.8, S.w.fg. = 0.8, S.w.fp. = 0.8, S.m.mg. = 0.9,
  S.m.mp. = 0.9, S.m.fg. = 0.9, S.m.fp. = 0.9, R.base.rate. = 1.8,
  R.hab.effect. = 0.5, co. = 1, f. = 0.5, S.b.mc. = 0.95,
  S.b.mk. = 0.85, S.b.md. = 0.8, S.b.fc. = 0.95, S.b.fk. = 0.85,
  S.f.mc. = 0.8, S.f.mk. = 0.75, S.f.md. = 0.8, S.f.fc. = 0.8,
  S.f.fk. = 0.75, S.y.mc. = 0.8, S.y.mk. = 0.75, S.y.fc. = 0.8,
  S.y.fk. = 0.75)
}
\arguments{
\item{scenario}{Difference scenarios defined by Runge & Marra: "winter" limited, "intermediate", and "spring" limited.  If set overides other varibles.}

\item{figure}{deprecated}

\item{gamma.}{xx}

\item{c.}{xxx}

\item{K.bc.}{xxx}

\item{K.bk.}{xxx}

\item{K.wg.}{xxx}

\item{S.w.mg.}{xxx}

\item{S.w.mp.}{xxx}

\item{S.w.fg.}{xxx}

\item{S.w.fp.}{xxx}

\item{S.m.mg.}{do they call this s.smg?}

\item{S.m.mp.}{xxx}

\item{S.m.fg.}{xxx}

\item{S.m.fp.}{xxx}

\item{R.base.rate.}{xxx}

\item{R.hab.effect.}{effects of pair breeding in poor habitat}

\item{co.}{magnitude of carry over effect; c 1 equals no carry over}

\item{f.}{sex ratio}

\item{S.b.mc.}{males in sour.c.e habitat}

\item{S.b.mk.}{males in sin.k. habitat}

\item{S.b.md.}{"drain" males}

\item{S.b.fc.}{xxx}

\item{S.b.fk.}{xxx}

\item{S.f.mc.}{males from sour.c.e}

\item{S.f.mk.}{xxx}

\item{S.f.md.}{drain males have higher surv b/c they don't have costs of repro}

\item{S.f.fc.}{xxx}

\item{S.f.fk.}{xxx}

\item{S.y.mc.}{xxx}

\item{S.y.mk.}{xxx}

\item{S.y.fc.}{xxx}

\item{S.y.fk.}{xxx}
}
\value{
param.ranges 30 x 2 dataframe of the min. and max. parameter values that will be considered.  Often values are fixed and do not vary.
}
\description{
Generates a dataframe defining the minimum maximum values of each parameter;
these are then turned into a matrix of all combinations by the function
makeParamCombos.df()
}
\examples{

# Default values
## Print out default values (only 1st 10 show.  Remove "[1:10]" to see all)
param_set()[1:10]

## Print defaults in a column by taking the transpose using t()
## (only 1st 10 show.  Remove "[1:10]" to see all)
t(param_set()[1:10])

# Set parameters for the different scenarios plotted in figure Figure 28.5 in Runge & Marra (2004)
## (note that to actually replicate Fig 28.5 param_rangs() and param_seqs() need to be used)
## to set up full parameter grid with the male dominanc parameter gamma varying.

## winter limited
winter <- param_set(scenario = "winter.limited")

## intermediate scenario
intermediate <- param_set(scenario = "intermediate")

## summer limited scenario
summer <- param_set(scenario = "summer.limited")

## compare the scenarios
### K.bc varies
### k.bk fixed
### K.wg varies
rbind(winter = winter, intermed = intermediate, summer = summer)[, 1:10]


}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds: The ecology & evolution of migration.
Johns Hopkins University Press, Baltimore.
}
