% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Fig28_3.R
\name{plot_Fig28_3}
\alias{plot_Fig28_3}
\title{Process the output of runMultiFAC() to re-create the 3D Fig 28.3}
\usage{
plot_Fig28_3(runFAC.multi, plot.type = "P", scale. = T, tick.labs = 1)
}
\description{
Process the output of runMultiFAC() to re-create the 3D Fig 28.3
}
