% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Fig28_4.R
\name{plot_Fig28_4}
\alias{plot_Fig28_4}
\title{Replicate plot 28.4}
\usage{
plot_Fig28_4(runFAC.multi, plot.debug.lines = T)
}
\arguments{
\item{runFAC.multi}{xxx}

\item{plot.debug.lines}{Plot lines which represent ...}
}
\description{
Functions to replicate plots that occur in Runge and Mara's Birds of Two Worlds Full Annual Cycle model
}
