% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Fig28_4_IBM.R
\name{plot_Fig28_4_IBM}
\alias{plot_Fig28_4_IBM}
\title{Replicate plot 28.4}
\usage{
plot_Fig28_4_IBM(runFAC.multi, plot.debug.lines = T)
}
\arguments{
\item{runFAC.multi}{Output of runFAC_multi}

\item{plot.debug.lines}{Should reference lines be plotted on graph}
}
\description{
Functions to replicate plots that occur in Runge and Mara's Birds of Two Worlds Full Annual Cycle model
}
