% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_finalize_output.R
\name{runFAC_finalize_output}
\alias{runFAC_finalize_output}
\title{Finalize output of a single run of FAC}
\usage{
runFAC_finalize_output(out.df, save.ts = TRUE)
}
\arguments{
\item{out.df}{Dataframe of output}

\item{save.ts}{Should the time series of population sizes be saved?}
}
\value{
out.df Dataframe ...
}
\description{
Finalize output of a single run of FAC
}
