% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_multi.R
\name{runFAC_multi}
\alias{runFAC_multi}
\title{Run multiple FAC models to equilibirum accross a range of parameters}
\usage{
runFAC_multi(param.grid = param_grid(), remakeFigure = NA)
}
\description{
Run multiple FAC models to equilibirum accross a range of parameters
}
