% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_multi_finalize_output.R
\name{runFAC_multi_finalize_output}
\alias{runFAC_multi_finalize_output}
\title{Finalize output from runFAC_multi()}
\usage{
runFAC_multi_finalize_output(multiFAC.out.df, param.grid, ...)
}
\arguments{
\item{multiFAC.out.df}{Output of runFAC_multi}

\item{param.grid}{Grid of parameters over which model was run (?)}

\item{...}{Additional arguements}
}
\value{
multiFAC.out.df ...
}
\description{
Finalize output from runFAC_multi()
}
