% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_store_pop_state_i.R
\name{runFAC_store_pop_state_i}
\alias{runFAC_store_pop_state_i}
\title{Save current population state}
\usage{
runFAC_store_pop_state_i(i, minimum.i = 10, FAC.out, W.list, B0.RM, B0.IB,
  P.breeding.pair.results.RM, P.breeding.pair.results.IB, Y2, A.G.i, A.P.i,
  check.eq = T, run.IBM = F, ...)
}
\arguments{
\item{i}{current iteration of model}

\item{FAC.out}{dataframe holding output of model for each iteraction}

\item{W.list}{list containing state of population after winter and fall migration}

\item{B0.RM}{vector of...}

\item{B0.IB}{vector of...}

\item{P.breeding.pair.results.RM}{vector representing pairing frequency of birds in source (".c") habitat the both originated from good (".gg") habitat on the wintering ground.}

\item{Y2}{vector of reproductive output (post migration?)}

\item{A.G.i}{vector...}

\item{A.P.i}{vector...}

\item{check.eq}{Should lambda be calculated}

\item{run.IBM}{}
}
\value{
FAC.out dataframe updated with population sizeds from time i
}
\description{
Save current population state
}
