% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_FAC_state.R
\name{save_FAC_state}
\alias{save_FAC_state}
\title{Save current population state}
\usage{
save_FAC_state(i, out.df, W.mg, W.mp, W.fg, W.fp, B0, P.cgg, P.cgp, P.cpg,
  P.cpp, P.kgg, P.kgp, P.kpg, P.kpp, Y2, A.G.i, A.P.i, ...)
}
\arguments{
\item{i}{current iteration of model}

\item{out.df}{dataframe holding output of model for each iteraction}

\item{W.mg}{scalar with number of males at beginning of winter in good habitat}

\item{W.mp}{scalar with number of males at beginning of winter in poor habitat}

\item{W.fg}{scalar with number of females at beginning of winter in good habitat}

\item{W.fp}{scalar with number of females at beginning of winter in poor habitat}

\item{B0}{vector of...}

\item{P.cgg}{scalar representing pairing frequency of birds in source (".c") habitat the both originated from good (".gg") habitat on the wintering ground.}

\item{P.cgp}{scalar...}

\item{P.cpp}{scalar...}

\item{P.kgg}{scalar...}

\item{P.kgp}{scalar...}

\item{P.kpg}{scalar...}

\item{P.kpp}{scalar...}

\item{Y2}{vector of reproductive output (post migration?)}

\item{A.G.i}{vector...}

\item{A.P.i}{vector...}
}
\value{
out.df dataframe updated with population sizeds from time i
}
\description{
Save current population state
}
