% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0_set_up.R
\name{step0_set_up}
\alias{step0_set_up}
\title{Step 0: Set up survival matrices &  other preliminaries}
\usage{
step0_set_up(param.set = param_set(), iterations = 150, use.IBM = F,
  use.IBM.S.b = F)
}
\arguments{
\item{param.set}{Initial parameters and population state for a single run of model to equilibirum}

\item{iterations}{Number of iterations to run model to allow it to reach equilibrium}

\item{use.IBM}{Use individual-based modeling implementation of breeding territory aquisition and mate pairing}

\item{use.IBM.S.b}{Used individual-based modeling implementation of breeding survibal.}
}
\description{
Step 0: Set up survival matrices &  other preliminaries
}
\examples{

# Initialize model
runFAC.i <- step0_set_up()

# Look at structure of list
## Elements for model output, matrices, parameters, etc
str(runFAC.i,1)

# Look at top of dataframe to hold output
head(runFAC.i$FAC.out.RM)[,1:5]

# Look at 2nd matrix used in model, winter survival
runFAC.i$param.matrices[2]
runFAC.i$param.matrices$S.w

}
