% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_winter_survival.R
\name{step1_winter_survival}
\alias{step1_winter_survival}
\title{FAC step 1: Execute 1st step of Runge & Marra FAC model: overwinter survival 1.RM and W1.IB, the pop stucture after overwinter survival}
\usage{
step1_winter_survival(
  i,
  runFAC.i,
  use.IBM,
  Ninit,
  W.list.RM = NULL,
  W.list.IB = NULL
)
}
\arguments{
\item{i}{xxx}

\item{runFAC.i}{xxx}

\item{use.IBM}{Use individual based model submodel}

\item{Ninit}{Initial population size}

\item{W.list.RM}{xxx}

\item{W.list.IB}{xxx}
}
\value{
W1.list
}
\description{
FAC step 1: Execute 1st step of Runge & Marra FAC model: overwinter survival 1.RM and W1.IB, the pop stucture after overwinter survival
}
\examples{

# Initial population size for iteration 1
## step1_winter_survival(i = 1, use.IBM = F,Ninit = c(10,0,10,0))
## step1_winter_survival(i = 1, use.IBM = T,Ninit = c(10,0,10,0))


}
