% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_substepA2_mate_acquire_IB.R
\name{step3_substepA2_mate_acquire_IB}
\alias{step3_substepA2_mate_acquire_IB}
\title{Step 3 Substep ...}
\usage{
step3_substepA2_mate_acquire_IB(hab.aq.results.IB)
}
\description{
Step 3 Substep ...
}
