% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_substepA2_mate_acquire_RM.R
\name{step3_substepA2_mate_acquire_RM}
\alias{step3_substepA2_mate_acquire_RM}
\title{Step 3 Substep A2}
\usage{
step3_substepA2_mate_acquire_RM(W2, param.set, hab.acquire.results)
}
\description{
Step 3 Substep A2
}
