% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_winter_habitat_acquisition.R
\name{step5_winter_habitat_acquisition}
\alias{step5_winter_habitat_acquisition}
\title{Execute 5th step of Runge & Mara FAC model: winter habitat acquisition}
\usage{
step5_winter_habitat_acquisition(B2, Y2, param.set, gamma.i, i, ...)
}
\arguments{
\item{B2}{Adult population state at end of fall migration}

\item{Y2}{Offspring population state at end of fall migration}

\item{param.set}{Initial parameters}

\item{gamma.i}{Vector of parameters defining relative competitive ability for winter habitat}

\item{i}{Model iteration number}

\item{...}{Additional parameters}
}
\description{
Execute 5th step of Runge & Mara FAC model: winter habitat acquisition
}
