% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substep_B_hab_acquire_RM.R
\name{substep_B_hab_acquire_RM}
\alias{substep_B_hab_acquire_RM}
\title{Breeding habitat acquisition substep of pre-breeding step of breeding season using original Runge & Marra allocation functions}
\usage{
substep_B_hab_acquire_RM(W2, K.bc = param.set$K.bc, K.bk = param.set$K.bk)
}
\arguments{
\item{W2}{population structure after winter and migration mortality}

\item{K.bc}{carrying capacity of breeding ground source (._c) habitat}

\item{K.bk}{carrying capacity (K) of breeding ground skin (._k) habitat}
}
\description{
For an alternative individual based (IB) approach see substep_breed_hab_acquire_IB()
}
