% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_P.R
\name{test_P}
\alias{test_P}
\title{QA/AC: Function to test equations 9 through 16}
\usage{
test_P(
  dat,
  FUN = eq09_Pcgg,
  plot.test = TRUE,
  subsample = TRUE,
  sample.N = 10,
  return.df = FALSE,
  ...
)
}
\arguments{
\item{dat}{dataframe with values against which to test function.  An object named test_dat_P is available for testing.}

\item{FUN}{function to test}

\item{plot.test}{plots a histogram of the proportions calculated by tyeh function}

\item{subsample}{should the test data be randomly subsampled for quick testing}

\item{sample.N}{how many subsamples of dataframe to take}

\item{return.df}{return the full dataframe with original test values}

\item{...}{Other}
}
\description{
These equations determine pairing proportions (P) in breeding habitat.
}
