% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_P_df_maker.R
\name{test_P_df_maker}
\alias{test_P_df_maker}
\title{make dataframe for test functions
to save use usethis::use_data(test.dat.P,overwrite = T)}
\usage{
test_P_df_maker(wmin = 1, wmax = 100, wstep = 30, kmin = 1,
  kmax = 125, kstep = 30)
}
\description{
make dataframe for test functions
to save use usethis::use_data(test.dat.P,overwrite = T)
}
