% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq01buildW0vect.R
\name{eq01buildW0vect}
\alias{eq01buildW0vect}
\title{Equation 1: Set up current population vector W0}
\usage{
eq01buildW0vect(W.mg, W.mp, W.fg, W.fp)
}
\arguments{
\item{W.mg}{scalar}

\item{W.mp}{scalar}

\item{W.fg}{scalar}

\item{W.fp}{scalar}
}
\value{
A vector containing the status of the popualtion at the beginning of winter: \code{W.mg} , \code{W.mp} , \code{W.fg} , and \code{W.fp}
}
\description{
Create a vector representing Population at beginning of winter (W0)
}
\details{
During 1st iteration (i=1) this taken from function call
This is updated during every iteration by taking values of W.mg, W.mp
etc that were calculated in equation 27 at the end of the last iteration (i-1)
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds.
Johns Hopkins University Press, Baltimore.
}
