% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq06_Bmc.R
\name{eq06_Bmc}
\alias{eq06_Bmc}
\title{Equation 6: Allocate breeding males to source habitat (B.mc)}
\usage{
eq06_Bmc(W2, K.bc)
}
\arguments{
\item{W2}{Population vector produced by eq02buildW1Mat()}

\item{K.bc}{Carrying capacity (K) during breeding season (b) of source (c) habitat}
}
\value{
B.mc Total number of males allocated to source habitat.
}
\description{
Equation 6: Allocate breeding males to source habitat (B.mc)
}
\details{
Summer male (M) dynamces
males acquiring  source habitat
I believe this allows males to be holding territories
w/o pairing w/ a female

Aliased as M.2.source.eq6.c()
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds.
Johns Hopkins University Press, Baltimore.
}
