% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq08_Bmd.R
\name{eq08_Bmd}
\alias{eq08_Bmd}
\title{Equation 8: Assigning breeding males to "drain" (aka "floater") status}
\usage{
eq08_Bmd(W2, K.bc, B.fk)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{source carrying capacity}

\item{B.fk}{females available in sink}
}
\value{
Number of males in drain habitat.
}
\description{
SUMMER MALE (M) DYNAMICS - drain males
Drain males: Males that don't find a territory AND don't find a mate (for those in sink)
become "drain" males (floaters)
mg+mp = total male population
minus those that ended up in source habitat (K.bc)
minus those that paired w/female in sink habitat (B.fk)
}
\details{
Note that it is possible for the equation W2.mg+W2.mp - K.bc to return a negative value if the number
of males is less than the source habitat size K.bc.  This is why  max() is used
to corred for this.
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds o
}
