% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq09_Pcgg.R
\name{eq09_Pcgg}
\alias{eq09_Pcgg}
\title{Equation 9: Pair formation in source habitat between birds which both originated from good winter habitat (P.cgg)}
\usage{
eq09_Pcgg(W2, K.bc, B.mc, B.fc, ...)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{source (c) carrying capacity (K) in breeding season (b)original file said "sink popualtion size" in comments}

\item{B.mc}{Number of males (.m) settled into source (.xc) habitat}

\item{B.fc}{females available in sink}

\item{...}{Aditional parameters}
}
\value{
P.cgg, the proportion of pairings in the "source" breeding habitat between male and females which both wintered in the good habitat
}
\description{
Calculates the proportion of pairs in source ("c") breeding habitat that occur between males that wintered in good habitat and females that wintered in good habitat ("gg").
}
\details{
The logic of this function is:
\enumerate{
\item IF \code{mg > K.bc}  (number of males from good habitat > source K)
AND \code{fg > K.bc} (number of females from good > source K)
THEN all of the pairs on the source habitat will be good-good pairs.
That is, if there are enough males and females to fill up the source then the good-good pairing rate will be 100\%. This will typically only occur when there are not many birds.
\item ELSE, the ratio of good-good pairs will be determined by the ratio of:
2a) The number of males OR females from good habitat, whichever is less
2b) AND the number of males or females on source habitat, whichever is less
}

B.fc should usually be lower than B.mc because females have lower survival and/or are forced by competition into poor habitat in the winter, resulting in carry over effects.

Possible ratios for the 2nd condition of the equation::
\enumerate{
\item W2.mg/B.mc = males from good winter / males settled in source
\item W2.fg/B.mc = females from good winter / males settled in source
\item W2.mg/B.fc = males from good winter / females settled in source
4)W2.fg/B.fc = females from good winter / females settled in source
}

The ratio is tricky.  It is determined by which sex is limiting in terms of input from good winter habitat (numerator) AND which sex is limiting in terms of total abundance in the breeding habitat.
}
\examples{

# Test eq09_Pcgg() for single sets of parameters
## Set parameters:
### Population size:
### Create named population W2 vector using eq01buildW0vect()
### Note: ratio of good male to good females is 10:1
W2 <- eq01buildW0vect(10,10,
                      1, 10)
### source Breeding habitat carrying capacities (K.bc)
#### 3 options
K.bc.10  <- 10
K.bc.1   <- 1
K.bc.0   <- 0

### Sink (k) Breeding habitat carrying capacity (K.bk)
K.bk <- 500

### Number of males and females allocated to source breeding habitat
#### use eq04_Bfc() & eq06_Bmc()

##### K.bc = 10
B.fc.K10 <- eq04_Bfc(W2, K.bc.10)
B.mc.K10 <- eq06_Bmc(W2, K.bc.10)

##### K.bc = 1
B.fc.K1  <- eq04_Bfc(W2, K.bc.1)
B.mc.K1  <- eq06_Bmc(W2, K.bc.1)

##### K.bc = 0
B.fc.K0  <- eq04_Bfc(W2, K.bc.0)
B.mc.K0  <- eq06_Bmc(W2, K.bc.0)
#'
#### K.bc = 10
##### 10 source territories, 10 good males, but only 1 good female
eq09_Pcgg(W2, K.bc = K.bc.10, B.fc = B.fc.K10,  B.mc = B.mc.K10)

#### K.bc = 1
##### 1 source territory, 10 good males, but only 1 good female
eq09_Pcgg(W2, K.bc = K.bc.1,  B.fc = B.fc.K1,   B.mc = B.mc.K1)

#### K.bc = 0
##### 0 source territories, 10 good males, but only 1 good female
eq09_Pcgg(W2, K.bc = K.bc.0,  B.fc = B.fc.K0,   B.mc = B.mc.K0)

# Test eq09_Pcgg() accross a range of parameter
dim(test_dat_P)
P.test.df.out <- test_P(test_dat_P, FUN = eq09_Pcgg, return.df = TRUE, plot.test = FALSE)

## Look at distribution of P.cgg
hist(P.test.df.out$P)
summary(P.test.df.out$P)
}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds.  Birds of two worlds. Johns Hopkins University Press, Baltimore.
}
