% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq12_Pcpp.R
\name{eq12_Pcpp}
\alias{eq12_Pcpp}
\title{Equation 12}
\usage{
eq12_Pcpp(P.cgg, P.cgp, P.cpg, ...)
}
\arguments{
\item{P.cgg}{proportion...}

\item{P.cgp}{proportion...}

\item{P.cpg}{proportion...}

\item{...}{....}
}
\value{
P.cpp
}
\description{
Equation 12
}
\details{
SUMMER MALE (M) DYNAMICS - pairing
proportion composed of a male and female both from poor habitat
This is calculated by subtraction

Note that in the original paper, equations 10 and 11 both are named "P.cgp" but
in equation 12 it is implied that 10 should be P.cgp while 11 should P.cpg.  THis
would mnake the 2nd subscript for the male condition (P.cg_) and the 3rd subscript
for the female condition (P.c_p)

Previously aliased as pairing.eq12.P.cpp()
}
