% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq12_Pcpp.R
\name{eq12_Pcpp}
\alias{eq12_Pcpp}
\title{Equation 12: Proportion of prairs between a male and female both from poor habitat (P.cpp)}
\usage{
eq12_Pcpp(P.cgg, P.cgp, P.cpg, ...)
}
\arguments{
\item{P.cgg}{Proportion of pairs in source (c) where male and female both originated from good (gg) winter habitat}

\item{P.cgp}{Proportion of pairs in source (c) where male originated from good (g) and female originated from poor winter habitat}

\item{P.cpg}{Proportion of pairs in source (c) where male originated from poor (p) and female originated from good (g) winter habitat}

\item{...}{Additional parameters}
}
\value{
P.cpp Proportion of prairs between a male and female both from poor habitat
}
\description{
Equation 12: Proportion of prairs between a male and female both from poor habitat (P.cpp)
}
\details{
SUMMER MALE (M) DYNAMICS - pairing
proportion composed of a male and female both from poor habitat
This is calculated by subtraction

Note that in the original paper, equations 10 and 11 both are named "P.cgp" but
in equation 12 it is implied that 10 should be P.cgp while 11 should P.cpg.  THis
would mnake the 2nd subscript for the male condition (P.cg_) and the 3rd subscript
for the female condition (P.c_p)

Previously aliased as pairing.eq12.P.cpp()
}
