% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq13_Pkgg.R
\name{eq13_Pkgg}
\alias{eq13_Pkgg}
\title{Equation 13: Pairing in sink (k) habitat of good-good pairs (P.kgg)}
\usage{
eq13_Pkgg(W2, K.bc, K.bk, B.mk, B.fk, ...)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{Carrying capacity during breeding season in source habitat}

\item{K.bk}{Carrying capacity during breeding season in sink habitat}

\item{B.mk}{Males (m) allocated to sink (k)}

\item{B.fk}{Females allocated to sink}

\item{...}{Additional parameters}
}
\value{
P.kgg Proportion of pairs in sink (k) habitat where both male and female originated from good winter habitat (gg)
}
\description{
Equation 13: Pairing in sink (k) habitat of good-good pairs (P.kgg)
}
\examples{
#Trivial example
W2. <- c(10,10,10,10)
names(W2.) <- c("mg","fg","mp","fp")
eq13_Pkgg(W2 = W2., K.bc = 0, K.bk = 10, B.mk = 5, B.fk =5)



}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds.  Birds of two worlds. Johns Hopkins University Press, Baltimore.
}
