% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq13_Pkgg.R
\name{eq13_Pkgg}
\alias{eq13_Pkgg}
\title{Equation 13: pairing in sink (.k) habitat}
\usage{
eq13_Pkgg(W2, K.bc, K.bk, B.mk, B.fk, i = NA)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{carrying capacity...}

\item{K.bk}{carrying capacity...}

\item{B.mk}{...}

\item{B.fk}{...}

\item{...}{}
}
\description{
Equation 13: pairing in sink (.k) habitat
}
\details{
good-good pairs that mate in SINK habitat

Previously aliased as pairing.eq13.P.kgg()
}
