% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq14_Pkgp.R
\name{eq14_Pkgp}
\alias{eq14_Pkgp}
\title{Equation 14: proportion in sink habitat, good-poor pairs}
\usage{
eq14_Pkgp(W2, K.bc, K.bk, B.mk, B.fk, ...)
}
\arguments{
\item{W2}{population vector}

\item{K.bc}{carrying capacity ...}

\item{K.bk}{carrying capacity ...}

\item{B.mk}{...}

\item{B.fk}{...}

\item{...}{...}
}
\description{
Equation 14: proportion in sink habitat, good-poor pairs
}
\details{
Previously aliased as pairing.eq14.P.kgp()
}
