% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq15_Pkpg.R
\name{eq15_Pkpg}
\alias{eq15_Pkpg}
\title{Equation 15: Proportion in sink poor-poor}
\usage{
eq15_Pkpg(W2, K.bc, K.bk, B.mk, B.fk, i = i)
}
\arguments{
\item{W2}{Popualtion vector}

\item{K.bc}{carrying capacity}

\item{K.bk}{carrying capacity}

\item{B.mk}{...}

\item{B.fk}{...}

\item{i}{index of loop.  for debugging}
}
\description{
Equation 15: Proportion in sink poor-poor
}
\details{
Aliased as pairing.eq15.P.kpg()
}
