% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq16_Pkpp.R
\name{eq16_Pkpp}
\alias{eq16_Pkpp}
\title{Equation 16:}
\usage{
eq16_Pkpp(P.kgg, P.kgp, P.kpg, B.mk, B.fk, ...)
}
\arguments{
\item{P.kgg}{xxx}

\item{P.kgp}{xxx}

\item{P.kpg}{xxx}

\item{B.mk}{xxx}

\item{B.fk}{xxx}

\item{...}{xxx}
}
\description{
Equation 16:
}
\details{
Proportion in sink composed of poor-poor

Previously aliased as pairing.eq16.P.kpp
}
