% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq19AbuildFmat.R
\name{eq19AbuildFmat}
\alias{eq19AbuildFmat}
\title{Equation 19: Matrix of sex ratios at birth (f.mat)}
\usage{
eq19AbuildFmat(f = 0.5)
}
\arguments{
\item{f}{Proportion of females}
}
\value{
f.mat matrix of sex ratios at bird
}
\description{
Equation 19: Matrix of sex ratios at birth (f.mat)
}
\examples{
# Sex ratio matrix
eq19AbuildFmat()
}
