% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq19BbuildMinMat.R
\name{eq19BbuildMinMat}
\alias{eq19BbuildMinMat}
\title{Equation 19: Build 2 x 2 Diagonal matrix based on minimnums of B.mc vs B.fc and B.mk vs. B.fk (min.mat)}
\usage{
eq19BbuildMinMat(B.mc, B.fc, B.mk, B.fk)
}
\arguments{
\item{B.mc}{Abundance of males ("m") in source ("c") habitat on the breeding ground}

\item{B.fc}{Females in source habitat}

\item{B.mk}{Abundance of males in sink ("k") habitat on the breeding ground}

\item{B.fk}{Females in sink habitat}
}
\value{
min.mat xxx
}
\description{
Equation 19: Build 2 x 2 Diagonal matrix based on minimnums of B.mc vs B.fc and B.mk vs. B.fk (min.mat)
}
