% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq20buildSbmat.R
\name{eq20buildSbmat}
\alias{eq20buildSbmat}
\title{Equation 20: Breeding season mortality matrix (S.b)}
\usage{
eq20buildSbmat(S.b.mc, S.b.mk, S.b.md, S.b.fc, S.b.fk)
}
\arguments{
\item{S.b.mc}{survival...}

\item{S.b.mk}{survival...}

\item{S.b.md}{survival...}

\item{S.b.fc}{survival...}

\item{S.b.fk}{survival...}
}
\value{
S.b Matrix of breed-ding season survival probabilities
}
\description{
Equation 20: Breeding season mortality matrix (S.b)
}
\details{
"adult birds experience both sex- and habitat-specific mortality
over the breeding season."
}
