% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq22buildSymat.R
\name{eq22buildSymat}
\alias{eq22buildSymat}
\title{Equation 22: Mortality during migration of young/offspring (S.y)}
\usage{
eq22buildSymat(S.y.mc, S.y.mk, S.y.fc, S.y.fk)
}
\arguments{
\item{S.y.mc}{survival...}

\item{S.y.mk}{survival...}

\item{S.y.fc}{survival...}

\item{S.y.fk}{survival...}
}
\description{
Equation 22: Mortality during migration of young/offspring (S.y)
}
\details{
Previously aliased as Fx.make.fall.adult.s.matrix.eq22
}
