% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq24_comp_loop.R
\name{eq24_comp_loop}
\alias{eq24_comp_loop}
\title{Competition loop}
\usage{
eq24_comp_loop(A.i.0, K.wg.0, y.i, comp.df, j, i, debug.continue = FALSE,
  internal.error.check = FALSE)
}
\description{
Competition loop
}
