% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq25_comp_constrain.R
\name{eq25_comp_constrain}
\alias{eq25_comp_constrain}
\title{Equation 25: Competition A.i constraint}
\usage{
eq25_comp_constrain(A.i.G.settled.raw.j, A.i.0)
}
\description{
Equation 25: Competition A.i constraint
}
