% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq26_alloc_winter_P.R
\name{eq26_alloc_winter_P}
\alias{eq26_alloc_winter_P}
\title{equation 26 Allocate to poor habitat}
\usage{
eq26_alloc_winter_P(A.i.0, A.i.G)
}
\arguments{
\item{A.i.0}{xxx}

\item{A.i.G}{xxx}
}
\description{
equation 26 Allocate to poor habitat
}
