% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq27_post_comp_pooling.R
\name{eq27_post_comp_pooling}
\alias{eq27_post_comp_pooling}
\title{Equation 27
Use of assign() here is probably inefficient/awk; put into list?}
\usage{
eq27_post_comp_pooling(A.G.i, A.P.i)
}
\description{
Equation 27
Use of assign() here is probably inefficient/awk; put into list?
}
