% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_build_Ptest_df.R
\name{error_check_build_Ptest_df}
\alias{error_check_build_Ptest_df}
\title{QA/QC: Build sequences of parameters for testing mate pairing equations}
\usage{
error_check_build_Ptest_df(
  include.W.0 = FALSE,
  include.K.0 = FALSE,
  seqW = seq(1, 100, 30),
  seqK = seq(1, 250, 50),
  seqP.cgg = NA,
  seqP.cgp = NA,
  seqP.cpg = NA,
  seqP.cpp = NA,
  seqP.kgg = NA,
  seqP.kgp = NA,
  seqP.kpg = NA,
  seqP.kpp = NA
)
}
\arguments{
\item{include.W.0}{xxx}

\item{include.K.0}{xxx}

\item{seqW}{xxx}

\item{seqK}{xxx}

\item{seqP.cgg}{xxx}

\item{seqP.cgp}{xxx}

\item{seqP.cpg}{xxx}

\item{seqP.cpp}{xxx}

\item{seqP.kgg}{xxx}

\item{seqP.kgp}{xxx}

\item{seqP.kpg}{xxx}

\item{seqP.kpp}{xxx}
}
\value{
P.test.df xxxx of parameters
}
\description{
QA/QC: Build sequences of parameters for testing mate pairing equations
}
\examples{
P.test.df <- error_check_build_Ptest_df()
dim(P.test.df)

}
