% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_P_test.R
\name{error_check_P_test}
\alias{error_check_P_test}
\title{Test mate pairing equation against test dataframe}
\usage{
error_check_P_test(
  equation = "eq9",
  P.test.df = error_check_build_Ptest_df(),
  print.i = FALSE
)
}
\arguments{
\item{equation}{Mate pairing equation to test.}

\item{P.test.df}{A dataframe}

\item{print.i}{print iteration number for debugging}
}
\description{
Test mate pairing equation against test dataframe
}
\examples{
P.test.df <- error_check_build_Ptest_df()
dim(P.test.df)

error_check_P_test(equation = "eq9",P.test.df=P.test.df[10,])
error_check_P_test(equation = "eq10",P.test.df=P.test.df[10,])

}
