% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_multiFAC_df.R
\name{make_multiFAC_df}
\alias{make_multiFAC_df}
\title{Make empty dataframe for holding output of runMultiFAC()}
\usage{
make_multiFAC_df(params.use = param_grid())
}
\arguments{
\item{params.use}{Dataframe of parameter values being used in the model.}
}
\description{
Make empty dataframe for holding output of runMultiFAC()
}
