% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_ranges.R
\name{param_ranges}
\alias{param_ranges}
\title{Make paramater ranges}
\usage{
param_ranges(scenario = NA, figure = NA, gamma. = c(5, 5), c. = c(1, 1),
  K.bc. = c(1000, 1000), K.bk. = c(10000, 10000), K.wg. = c(900, 900),
  S.w.mg. = c(0.8, 0.8), S.w.mp. = c(0.8, 0.8), S.w.fg. = c(0.8, 0.8),
  S.w.fp. = c(0.8, 0.8), S.m.mg. = c(0.9, 0.9), S.m.mp. = c(0.8, 0.8),
  S.m.fg. = c(0.9, 0.9), S.m.fp. = c(0.8, 0.8), R.base.rate. = c(1.8,
  1.8), R.hab.effect. = c(0.5, 0.5), co. = c(1, 1), f. = c(0.5, 0.5),
  S.b.mc. = c(0.95, 0.95), S.b.mk. = c(0.85, 0.85), S.b.md. = c(0.8, 0.8),
  S.b.fc. = c(0.95, 0.95), S.b.fk. = c(0.85, 0.85), S.f.mc. = c(0.8, 0.8),
  S.f.mk. = c(0.75, 0.75), S.f.md. = c(0.8, 0.8), S.f.fc. = c(0.8, 0.8),
  S.f.fk. = c(0.75, 0.75), S.y.mc. = c(0.8, 0.8), S.y.mk. = c(0.75, 0.75),
  S.y.fc. = c(0.8, 0.8), S.y.fk. = c(0.75, 0.75), verbose = TRUE)
}
\arguments{
\item{scenario}{Difference scenarios defined by Runge & Marra.  If set overides other varibles that may have been set using other arguements.}

\item{figure}{generate figure.  Currently only figures 28.3 and 28.4 are implemented}

\item{gamma.}{xx}

\item{c.}{xxx}

\item{K.bc.}{xxx}

\item{K.bk.}{xxx}

\item{K.wg.}{xxx}

\item{S.w.mg.}{xxx}

\item{S.w.mp.}{xxx}

\item{S.w.fg.}{xxx}

\item{S.w.fp.}{xxx}

\item{S.m.mg.}{do they call this s.smg?}

\item{S.m.mp.}{xxx}

\item{S.m.fg.}{xxx}

\item{S.m.fp.}{xxx}

\item{R.base.rate.}{xxx}

\item{R.hab.effect.}{effects of pair breeding in poor habitat}

\item{co.}{magnitude of carry over effect; c 1 equals no carry over}

\item{f.}{sex ratio}

\item{S.b.mc.}{males in sour.c.e habitat}

\item{S.b.mk.}{males in sin.k. habitat}

\item{S.b.md.}{"drain" males}

\item{S.b.fc.}{xxx}

\item{S.b.fk.}{xxx}

\item{S.f.mc.}{males from sour.c.e}

\item{S.f.mk.}{xxx}

\item{S.f.md.}{drain males have higher surv b/c they don't have costs of repro}

\item{S.f.fc.}{xxx}

\item{S.f.fk.}{xxx}

\item{S.y.mc.}{xxx}

\item{S.y.mk.}{xxx}

\item{S.y.fc.}{xxx}

\item{S.y.fk.}{xxx}

\item{verbose}{T/F return messages}
}
\value{
param.ranges 30 x 2 dataframe of the min. and max. parameter values that will be considered.  Often values are fixed and do not vary.
}
\description{
Generates a dataframe defining the minimum maximum values of each parameter;
these are then turned into a matrix of all combinations by the function
makeParamCombos.df()
}
\examples{

# Set parameters for the winter limited scenario in original paper (Runge & Marra 2004)
param.ranges <- param_ranges(scenario = "winter.limited")
head(param.ranges)

# Parameters to remake Figure 28.3 in original paper (Runge & Marra 2004)
param.ranges <- param_ranges(figure = 28.3)
param.ranges[c("K.bc","K.wg"), ]

# Setting carrying capacities to 0 causes errors when running model
## param_ranges() therefore automatically changes to prevent this bug
param.ranges <- param_ranges(K.bc. = c(0,100))
param.ranges[c("K.bc","K.wg"), ]

# Set up different scenarios
## this is similar to data for figure 28.5, but without gamma varying
winter.lim <- param_ranges(scenario = "winter")
intermed <- param_ranges(scenario = "intermediate")
summer.lim <- param_ranges(scenario = "summer")

# Show data
## Note only 1st 6 rows shown; delete "[1:6,]" to show all.
data.frame(winter.lim, intermed, summer.lim)[1:6,]

}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds: The ecology & evolution of migration.
Johns Hopkins University Press, Baltimore.
}
