% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_seqs.R, R/param_seqs2.R
\name{param_seqs}
\alias{param_seqs}
\title{Create sequence of parameter values between minimum and maximum.  Output is a list}
\usage{
param_seqs(param.ranges = param_ranges(), length.out = 10)

param_seqs(param.ranges = param_ranges(), length.out = 10)
}
\arguments{
\item{param.ranges}{minimum and maximum values for all parameters}

\item{length.out}{number of increments to divide range into}

\item{len.out}{number of increments to divide range into}

\item{param.ranges}{minimum and maximum values for all parameters, generted by \link{param_range}}
}
\value{
param.seq List with 30 elements, one for each parameter.

param.seq List with 30 elements, one for each parameter.
}
\description{
Create sequence of parameter values between minimum and maximum.  Output is a list

Create sequence of parameter values between minimum and maximum.  Output is a list
}
\examples{
#generate min and max values for each parameter used for figure 28.3 in original paper
params <- param_ranges(figure = 28.3)

# Generate parameter sequence
params.seqs <- param_seqs(param.ranges = params)


## Most are constants; 3rd and 5th element vary accross a range of values
params.seqs$K.wg



}
