% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{param_set}
\alias{param_set}
\title{Set individual parameter values.}
\usage{
param_set(scenario = NA, figure = NA, gamma. = 5, c. = 1,
  K.bc. = 1000, K.bk. = 10000, K.wg. = 900, S.w.mg. = 0.8,
  S.w.mp. = 0.8, S.w.fg. = 0.8, S.w.fp. = 0.8, S.m.mg. = 0.9,
  S.m.mp. = 0.9, S.m.fg. = 0.9, S.m.fp. = 0.9, R.base.rate. = 1.8,
  R.hab.effect. = 0.5, co. = 1, f. = 0.5, S.b.mc. = 0.95,
  S.b.mk. = 0.85, S.b.md. = 0.8, S.b.fc. = 0.95, S.b.fk. = 0.85,
  S.f.mc. = 0.8, S.f.mk. = 0.75, S.f.md. = 0.8, S.f.fc. = 0.8,
  S.f.fk. = 0.75, S.y.mc. = 0.8, S.y.mk. = 0.75, S.y.fc. = 0.8,
  S.y.fk. = 0.75)
}
\arguments{
\item{scenario}{Different habitat scenarios defined by Runge & Marra 2004.  In particular sets carrying capacity for breeding source habitat (K.bc) and winter good habitat (K.wg)    If set overides other varibles that may have been set using other arguements.}

\item{figure}{deprecated.  See param_ranges()}

\item{gamma.}{Strength of male dominance.}

\item{c.}{Strength of winter to breeding season carry over.}

\item{K.bc.}{Carrying capacity (K) of source (high quality) breeding habitat.}

\item{K.bk.}{Carrying capacity (K) of sink (low quality) breeding habitat.}

\item{K.wg.}{Carrying capacity (K) of winter good (high quality) habitat.}

\item{S.w.mg.}{Survival (S) during winter (.w.) of males (.m) in good habitat (g).  Note that there is no age structure in winter.}

\item{S.w.mp.}{Survival (S) during winter (.w.) of males (.m) in poor habitat (p).}

\item{S.w.fg.}{Survival (S) during winter (.w.) of females (.f) in good habitat (g).}

\item{S.w.fp.}{Survival (S) during winter (.w.) of females (.f) in good habitat (p).}

\item{S.m.mg.}{Survival (S) during spring migration (.m.) of males (.m) originating from good (g) winter habitat.}

\item{S.m.mp.}{Survival (S) during spring migration (.m.) of males (.m) originating from poor (p) winter habitat.}

\item{S.m.fg.}{Survival (S) during spring migration (.m.) of females (.f) originating from good (g) winter habitat.}

\item{S.m.fp.}{Survival (S) during spring migration (.m.) of females (.m) originating from poor (p) winter habitat.}

\item{R.base.rate.}{Fecundity base rate (?)}

\item{R.hab.effect.}{Effect on fecundity of a pair breeding in poor habitat}

\item{co.}{Magnitude of carry over effect; c = 1 is no carry over}

\item{f.}{Sex ratio}

\item{S.b.mc.}{Survival (S) during breeding (.b.) of males (.m) in source (c) habitat (source = good habitat)}

\item{S.b.mk.}{Survival (S) during breeding (.b.) of males (.m) in sink (k) habitat (sink = poor habitat)}

\item{S.b.md.}{Survival (S) during breeding (.b.) of males (.m) with drain (d) status (drain = floater or unpaired)}

\item{S.b.fc.}{Survival (S) during breeding (.b.) of females (.f) in source (c) habitat (source = good habitat)}

\item{S.b.fk.}{Survival (S) during breeding (.b.) of females (.f) in sink (k) habitat (sink = poor habitat)}

\item{S.f.mc.}{Survival (S) during fall (.f.) migration of males (.m) originating from source (c) habitat}

\item{S.f.mk.}{Survival (S) during fall (.f.) migration of males (.m) originating from sink (k) habitat}

\item{S.f.md.}{Survival (S) during fall (.f.) migration of males (.m) with drain (d) status (aka floater, unpaired)}

\item{S.f.fc.}{Survival (S) during fall (.f.) migration of females (.f) originating from source (c) habitat}

\item{S.f.fk.}{Survival (S) during fall (.f.) migration of females (.f) originating from sink (k) habitat}

\item{S.y.mc.}{Survival (S) during fall migration of young (.y.) males (.m) originating from source (c) habitat}

\item{S.y.mk.}{Survival (S) during fall migration of young (.y.) males (.m) originating from sink (k) habitat}

\item{S.y.fc.}{Survival (S) during fall migration of young (.y.) females (.f) originating from source (c) habitat}

\item{S.y.fk.}{Survival (S) during fall migration of young (.y.) females (.f) originating from sink (k) habitat}
}
\value{
param.set 30 x 1 dataframe of theparameter values that will be used.
}
\description{
Sets individual parameter values for a single run of the model with runFAC().
}
\details{
See param_ranges() for setting the min and max of parameters that vary when using runFAC_multi
}
\examples{

# Default values
## Print out default values (only 1st 10 show.  Remove "[1:10]" to see all)
param_set()[1:10]

## Print defaults in a column by taking the transpose using t()
## (only 1st 10 show.  Remove "[1:10]" to see all)
t(param_set()[1:10])

# Set parameters for the different scenarios plotted in figure Figure 28.5 in Runge & Marra (2004)
## (note that to actually replicate Fig 28.5 param_rangs() and param_seqs() need to be used)
## to set up full parameter grid with the male dominanc parameter gamma varying.

## winter limited
winter <- param_set(scenario = "winter.limited")

## intermediate scenario
intermediate <- param_set(scenario = "intermediate")

## summer limited scenario
summer <- param_set(scenario = "summer.limited")

## compare the scenarios
### K.bc varies
### k.bk fixed
### K.wg varies
rbind(winter = winter, intermed = intermediate, summer = summer)[, 1:10]


}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds: The ecology & evolution of migration.
Johns Hopkins University Press, Baltimore.
}
