% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Fig28_4.R
\name{plot_Fig28_4}
\alias{plot_Fig28_4}
\title{Replicate line graph Plot 28.4 showing changes in age structure as carrying capacity changes.}
\usage{
plot_Fig28_4(runFAC.multi, plot.debug.lines = F)
}
\arguments{
\item{runFAC.multi}{Output draframe from runFAC.multi()}

\item{plot.debug.lines}{Plot lines to faciltate comparison of redstart output to the original Figure 28.4.}
}
\description{
Functions to replicate plots that occur in Runge and Mara's Birds of Two Worlds Full Annual Cycle model
}
