% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Fig28_5.R
\name{plot_Fig28_5}
\alias{plot_Fig28_5}
\title{Replicate line graph Figure 28.5 varying male dominance parameter gamma}
\usage{
plot_Fig28_5(dat.winter.lim, dat.intermediate, dat.summer.lim, drain = T)
}
\arguments{
\item{dat.winter.lim}{Dataframe from running a winter-limited FAC}

\item{dat.intermediate}{Dataframe representing an intermediate case between winter and summer limitation}

\item{dat.summer.lim}{Dataframe from running a summer-limited FAC}

\item{drain}{Should drain males be included in plots.}
}
\description{
This varies competition for high quality winter habitat
}
