% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_runFAC.R
\name{plot_runFAC}
\alias{plot_runFAC}
\title{Plot time series of a single run of a FAC model}
\usage{
plot_runFAC(out.df, y1 = "tot.W", y2 = "tot.B", ...)
}
\description{
Plot time series of a single run of a FAC model
}
