% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC.R
\name{runFAC}
\alias{runFAC}
\title{Run FAC model to equilibirum}
\usage{
runFAC(iterations = 350, Ninit = c(10, 0, 10, 0), param.set = param_set(),
  scenario = NA, vary = NA, verbose = FALSE, check.errors.in = c("B0",
  "P.cgg", "P.cgp", "P.cpg", "P.cpp", "P.kgg", "P.kgp", "P.kpg", "P.kpp", "Y1"),
  check.eq = TRUE, check.eq.after.i = 20, tol.1 = 3, tol.2 = 3,
  diagnostic.plot = T, return.output = T, save.ts = T, ...)
}
\arguments{
\item{iterations}{how many iterations of model to run before stopping}

\item{Ninit}{vector of initial abundances at the begining of winter for W.mg, W.mp, W.fg and W.fp}

\item{param.set}{initila parameters for this run of the model.}

\item{scenario}{character string representing name of scenario explored in original Runge and Marra paper (not currently implemented 7/10/2018)}

\item{vary}{...}

\item{verbose}{verbose general output}

\item{check.errors.in}{run error checks and report}

\item{check.eq}{should equilibrium be assessed?}

\item{tol.1}{tolerance value for diagnostics checks for whether the run of the model has reached equilibrium.  Lower values are more stringent; minimum is 1, which makes the model run the full number of iterations}

\item{tol.2}{secondary tolerance value for checking equilbrium; LARGER values increase stringency}

\item{diagnostic.plot}{Return a diagnostic plot show population size over time}

\item{return.output}{Output the full model dataframe?}

\item{save.ts}{Save the full time series of the model run?  If FALSE then only final time point will be returned.  Additionally, equilibrium monitoring will not be done and the time series cannot be plotted}

\item{...}{other arguements passed}

\item{minimum}{number of iterations to run if equilibrium is being checked}
}
\value{
A dataframe containing the status of the popualtion at each time step for all parameters
}
\description{
Iterates a single parameterization of the Runge & Marra 2004 FAC model
until equilbirium is reached.  Typically the desired output is the equilibrium
population size(s) and all intermediate output is discarded.
}
\references{
Runge, MC & PP Marra 2004.  Modeling seasonal interactions in the population dynamics of migratory birds.  Birds of two words.
}
