% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC.R
\name{runFAC}
\alias{runFAC}
\title{Run a single FAC model to equilibirum}
\usage{
runFAC(
  iterations = 350,
  Ninit = c(10, 0, 10, 0),
  param.set = param_set(),
  scenario = NA,
  vary = NA,
  verbose = FALSE,
  check.errors = TRUE,
  check.errors.in = c("B0.RM", "P.cgg", "P.cgp", "P.cpg", "P.cpp", "P.kgg", "P.kgp",
    "P.kpg", "P.kpp", "Y1"),
  check.eq = TRUE,
  minimum.i = 20,
  eq.tol = 6,
  diagnostic.plot = F,
  return.output = T,
  save.ts = T,
  use.IBM = F,
  use.IBM.S.b = F,
  ...
)
}
\arguments{
\item{iterations}{how many iterations of model to run before stopping}

\item{Ninit}{vector of initial abundances at the begining of winter for W.mg, W.mp, W.fg and W.fp}

\item{param.set}{initila parameters for this run of the model.}

\item{scenario}{character string representing name of scenario explored in original Runge and Marra paper (not currently implemented 7/10/2018)}

\item{vary}{...}

\item{verbose}{verbose general output}

\item{check.errors}{Run error checks on model subcomponents}

\item{check.errors.in}{run error checks and report}

\item{check.eq}{should equilibrium be assessed?}

\item{minimum.i}{number of iterations to run if equilibrium is being checked}

\item{eq.tol}{Tolerance for equilibirum check; the number of digistic lambda and variace of lambda should be rounded to when comparing against 1.  Larger numbers reduce the likelihood of model passing the equilbirium check}

\item{diagnostic.plot}{Return a diagnostic plot show population size over time}

\item{return.output}{Output the full model dataframe?}

\item{save.ts}{Save the full time series of the model run?  If FALSE then only final time point will be returned.  Additionally, equilibrium monitoring will not be done and the time series cannot be plotted}

\item{use.IBM}{use individual based model for habitat and mate acquisition}

\item{use.IBM.S.b}{use IBM for summer adult survival}

\item{...}{other arguements passed}
}
\value{
runFAC.i A dataframe containing the status of the population at each time step for all parameters
}
\description{
Iterates a single parameterization of the Runge & Marra 2004 FAC model
until equilbirium is reached.  Typically the desired output is the equilibrium
population size(s) and all intermediate output is discarded.
}
\examples{
# Run an FAC
test.FAC <- runFAC(verbose = TRUE)

# Look at structure of output
str(test.FAC,1)

}
\references{
Runge, MC and PP Marra.  2004.  Modeling seasonal
interactions in the population dynamics of migratory birds.
In Greenberg, R and PP Marra, eds.  Birds of two worlds.
Johns Hopkins University Press, Baltimore.
}
