% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_check_equilibrium.R
\name{runFAC_check_equilibrium}
\alias{runFAC_check_equilibrium}
\title{check model equilibrium}
\usage{
runFAC_check_equilibrium(lamb.B.mc.mean, lamb.B.mc.var, i, at.eq = FALSE,
  verbose = TRUE, eq.tol = 5, ...)
}
\description{
check model equilibrium
}
