% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_check_equilibrium.R
\name{runFAC_check_equilibrium}
\alias{runFAC_check_equilibrium}
\title{check model equilibrium}
\usage{
runFAC_check_equilibrium(lamb.B.mc.mean, lamb.B.mc.var, i, at.eq = FALSE,
  verbose = TRUE, eq.tol = 5, ...)
}
\arguments{
\item{lamb.B.mc.mean}{Mean lambda from B.mc ..}

\item{lamb.B.mc.var}{Variance of lambda from B.mc ...}

\item{i}{iteration}

\item{at.eq}{At equilibrium}

\item{verbose}{Verbose output?}

\item{eq.tol}{Equilibrium tolerance level (higher is more conservative)}
}
\description{
check model equilibrium
}
