% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_multi.R
\name{runFAC_multi}
\alias{runFAC_multi}
\title{Run multiple FAC models to equilibirum accross a range of parameters}
\usage{
runFAC_multi(param.grid = param_grid(), remakeFigure = NA)
}
\arguments{
\item{para.grid}{Dataframe of parameters, usually with 1 or 2 that vary accross a range}
}
\description{
Note that the "para.grid" arguement take take produced by the param_grid() function,
which gets its name because it uses the expand.grid() function to make all
combinations of the parameters being varied.
runFAC() has been updated to be able to output the full time serie
OR just the eq pop size
need to update runFAC_multi() to just take that eq size
should run a bit faster if runFAC() is keep all of that info
NB: "remakeFigure" does not appear in current code
}
