% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFAC_multi_finalize_output.R
\name{runFAC_multi_finalize_output}
\alias{runFAC_multi_finalize_output}
\title{Finalize output from runFAC_multi}
\usage{
runFAC_multi_finalize_output(multiFAC.out.df, param.grid)
}
\description{
Finalize output from runFAC_multi
}
