% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step0_set_up.R
\name{step0_set_up}
\alias{step0_set_up}
\title{Step 0: Set up survival matrices and other preliminaries}
\usage{
step0_set_up(param.set = param_set(), iterations = 150, use.IBM = F,
  use.IBM.S.b = F)
}
\description{
Step 0: Set up survival matrices and other preliminaries
}
\examples{

# Generate default matrices
runFAC.i <- runFAC_set_initial_params()

# Look at structure of list
str(runFAC.i)

# Look at top of dataframe to hold output
head(runFAC.i$FAC.out)[,1:5]

# Look at 2nd matrix, winter survival
runFAC.i$param.matrices[2]
runFAC.i$param.matrices$S.w
}
