% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_winter_survival.R
\name{step1_winter_survival}
\alias{step1_winter_survival}
\title{Execute 1st step of Runge & Marra FAC model: overwinter survival}
\usage{
step1_winter_survival(i, runFAC.i, use.IBM, Ninit, W.list.RM = NULL,
  W.list.IB = NULL)
}
\arguments{
\item{i}{model iterations}

\item{runFAC.i}{Initial parameters for this run of the FAC}

\item{use.IBM}{Use individual-based modeling approach instead of Runge & Marra equation (not currently implemented)}

\item{Ninit}{vector of initial population sizes for 1st iteration (i =1)}

\item{W.list.RM}{List containing post-competition sub-population sizes for Runge-Marra implementation.  Takes place of Ninit on all iterations > 1}
}
\value{
W1.list list containing W1.RM and W1.IB, the population stucture after overwinter survival
}
\description{
Execute 1st step of Runge & Marra FAC model: overwinter survival
}
\examples{
W.list.IB List containing post-competition sub-population sizes for individual based model implementation (IBM).  Takes place of Ninit on all iterations > 1

# Initial population size for iteration 1
step_1_winter_survival(i = 1, use.IBM = F,Ninit = c(10,0,10,0))
step_1_winter_survival(i = 1, use.IBM = T,Ninit = c(10,0,10,0))


}
