% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_substepA1_hab_acquire_IB.R
\name{step3_substepA1_hab_acquire_IB}
\alias{step3_substepA1_hab_acquire_IB}
\title{FAC-IBM Step 3, Substep A1: Breeding habitat acquisition substep of pre-breeding step of breeding season using individual-based (IB) approach}
\usage{
step3_substepA1_hab_acquire_IB(W2, K.bc, K.bk, ...)
}
\arguments{
\item{W2}{...}

\item{K.bc}{...}

\item{K.bk}{...}

\item{...}{additional arguements}
}
\value{
hab.aq.results.IB Results of habitatio acquisition
}
\description{
FAC-IBM Step 3, Substep A1: Breeding habitat acquisition substep of pre-breeding step of breeding season using individual-based (IB) approach
}
\examples{
#Set up param.set
param.set <- redstart::param_set()

#set breeding carrying capacities to be relatively small
param.set$K.bc <- 50
param.set$K.bk <- 50

#Set up W2 vector
W2 <- c(10,5,10,10); names(W2) <- c("mg","mp","fg","fp")


out <- step3_substepA1_hab_acquire_IB(W2 = W2,
           K.bc =  param.set$K.bc,
           K.bk = param.set$K.bk)

summary(out)

}
