% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_substepA2_mate_acquire_IB.R
\name{step3_substepA2_mate_acquire_IB}
\alias{step3_substepA2_mate_acquire_IB}
\title{FAC-IBM Step 3, Substep A2: Mate acquition}
\usage{
step3_substepA2_mate_acquire_IB(hab.aq.results.IB)
}
\arguments{
\item{hab.aq.results.IB}{Output from individuals-based model}
}
\value{
P.breeding.pair.results.list.IB
}
\description{
FAC-IBM Step 3, Substep A2: Mate acquition
}
