% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_P.R
\name{test_P}
\alias{test_P}
\title{Function to test equations 9 through 16}
\usage{
test_P(dat = test.dat.P, FUN = eq09_Pcgg, plot.test = TRUE,
  subsample = TRUE, sample.N = 10, return.df = FALSE, ...)
}
\arguments{
\item{FUN}{function to test}

\item{plot.test}{plots a histogram of the proportions calculated by tyeh function}

\item{subsample}{should the test data be randomly subsampled for quick testing}

\item{sample.N}{how many subsamples of datarame to take}

\item{return.df}{return the full dataframe with original test values}

\item{test.dat.P}{dataframe with values against which to test function}
}
\description{
These equations determine pairing proportions (P) in breeding habitat.
}
